Fix broken safehtml template setting
With the release of Hugo 0.60 a new markdown renderer was introduced.
The new renderer broke the `{{ .Content | safehtml }}` setting by not rendering inline HTML into the output (although it is explicitly allowed by Markdown)

To fix this a new global setting has been introduced, which is now enabled by this commit.